cut+paste new file
